fix: Refactor Qubit and QubitSet to a separate module
