+++ src/braket/circuits/unitary_calculation.py
@@ -19,8 +19,8 @@ from scipy.linalg import fractional_matrix_power
 from braket.circuits.compiler_directive import CompilerDirective
 from braket.circuits.gate import Gate
 from braket.circuits.instruction import Instruction
-from braket.circuits.qubit_set import QubitSet
 from braket.default_simulator.linalg_utils import multiply_matrix
+from braket.qubits.qubit_set import QubitSet
 
 
 def _einsum_subscripts(targets: QubitSet, qubit_count: int) -> str:
