+++ src/braket/qubits/__init__.py
@@ -11,4 +11,5 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from braket.qubits import Qubit, QubitInput  # noqa: F401
+from braket.qubits.qubit import Qubit, QubitInput  # noqa: F401
+from braket.qubits.qubit_set import QubitSet, QubitSetInput  # noqa: F401
