+++ test/unit_tests/braket/pulse/ast/test_approximation_parser.py
@@ -19,12 +19,12 @@ import pytest
 from openpulse import ast
 from oqpy import IntVar
 
-from braket.circuits.qubit_set import QubitSet
 from braket.pulse import ArbitraryWaveform, ConstantWaveform, DragGaussianWaveform, GaussianWaveform
 from braket.pulse.ast.approximation_parser import _ApproximationParser
 from braket.pulse.frame import Frame
 from braket.pulse.port import Port
 from braket.pulse.pulse_sequence import PulseSequence
+from braket.qubits.qubit_set import QubitSet
 from braket.timings.time_series import TimeSeries, _all_close
