+++ test/unit_tests/braket/qubits/test_qubit_set.py
@@ -13,7 +13,7 @@
 
 import pytest
 
-from braket.circuits import Qubit, QubitSet
+from braket.qubits import Qubit, QubitSet
 
 
 @pytest.fixture
